backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
mysqlclient>=2.2.0
python-multipart>=0.0.5
passlib[bcrypt]>=1.7.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pydantic>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
openai>=1.0.0
gitpython>=3.1.0
docker>=6.0.0